Experimental

+++ LearnGL/CustomSoundFormat.hpp
@@ -9,7 +9,7 @@
 #pragma pack(push,1)
 struct Four_bit {
 
-	int8_t value : 4;
+	uint8_t value : 4;
 
 };
 #pragma pack(pop)
@@ -29,7 +29,7 @@ struct Block {
 struct WL
 {
 	uint8_t magic;
-	uint16_t block_count;
+	uint32_t block_count;
 	uint16_t sample_rate;
 
 	Block* data;
